Add rehash and restart commands, edit ch
The bot needed a rehash/restart function. As it needed such a
functionality as 'rehash', I've also added a function to the main core
to allow reloading the settings (self.settingsdb) variable. Thus, we'll
call this whenever we need to rehash the settings. This is great if
there's been a change to the Redis database after the bot started. That
way, we will not have to reboot the bot in the event the Redis database
needs to be edited at any given time. This includes adding people to the
accesslist.

Signed-off-by: Dave Russell Jr <dave@createazure.com>