Added ability to stop/start the bot, and ack event
This commit introduces the ability to start and stop the events timer.
Although the start switch isn't quite finished yet, I'm still ironing
that out. APScheduler doesn't like me.

This also introduces the ability to acknowledge events, and I've
modified confidence's core to allow for automatic event sauce. The
current interval is every two minutes, but I will likely add a rule
where events will be cached and only displayed if they haven't been
displayed for X minutes... Currently it floods the IRC channel.

There's also a known bug where the bot will quit out of the blue... I
think that's because of the amount of time it pings, though. I think I'm
going to rate limit the amount of requests. Twisted framework lets you
do that, but it doesn't work ideally because it actually limits
everything that's being sent to the server per second. I may use the
Twisted framework to do so, but I doubt it... I didn't like the results
I got with it before.

Anyways, here goes nothing!

Signed-off-by: Dave Russell Jr <dave@createazure.com>