Fix broken arise command, add redis to ack/events
This commit will fix the 'arise' command, which was broken. I've moved
the entire functionality of apscheduler into its own method within the
core of the bot, and that seems to have fixed it. Therefore, the arise
command basically just calls that method and self.msg() to the logging
channel.

Redis will now have an effect on acknowledgements and events as they're
displayed in the logging channel. After the event is displayed in the
logging channel, it will add it to the redis database as
triggerid-hostid, concatenated as a string. The value of this key will
be event['lastchange']. If that timestamp changes, the bot will update
the channel again.

Keeping in mind that anything that's not acknowledged will be placed in
the channel, this will enforce acknowledging things.

Signed-off-by: Dave Russell Jr <dave@createazure.com>