Access list checking made global
Originally, the only command that used access list checking was restart,
but this will now be made global throughout all of the commands in
commandhandler.py.

The Redis database will take into account the hostname of the user as
well, so if the host name does not match, they're out of luck. This is
for added security (it's harder to spoof a hostname on IRC than a
username).

Signed-off-by: Dave Russell Jr <dave@createazure.com>